refactor(kagent): move deployment finder into the kagent package
NewStoreDeploymentFinder takes a Deployment store directly; the generic
database package no longer carries Kagent-specific lookup code.
